some PR feedback
