+++ workers/unity/Packages/com.improbable.gdk.mobile/Editor/AndroidLaunchMenu.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4768e3f66fbbc374d82396f8390d7b50
+guid: 1b81b5388a2444fa8b067a2685219b82
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
