+++ workers/unity/Packages/com.improbable.gdk.mobile/Editor/MobileLaunchHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a8bda4041db514fca958e3bba7b97966
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
