+++ workers/unity/Packages/com.improbable.gdk.mobile/Editor/iOSLauchMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e9c55ced9657940dd9f60c06a509ca91
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
